Working on button.
For now I have a working version of the order button. There are several
break cases though including changing screen size and using .length has
different numbers based on the rgb value. This leads to the order
button ‘activating’ when only five colors are picked if the colors have
over 200 for each R G B value and the button not activating if the
values are under 10 (all white or all black are good examples of colors
that will break my current code).