Modal Form Dialog Added
I added jQuery UI to the project for the modal form function. I may
need to reevaluate my decision based on load time but for now I am
happy with the functionality added. At this point the form is just a
filler and has a problem of continually opening an alert over and over.