Finishing touches on design.
I added finishing touches on the design like a hover state to the order
button when it is activated and pointer cursors in the color selection
boxes. Also the dialog while not responsive to screen size changes on
the fly should open on any screen size and look good. I chose this
option because it added 10 lines of code over one that would have added
200 lines of code but changed size as the window was resized.